あ

+++ Stage.cpp
@@ -108,7 +108,7 @@ BOOL Stage::DialogProc(HWND hDlg, UINT msg, WPARAM wp, LPARAM lp)
 	case WM_COMMAND:
 		mode_ = LOWORD(wp) - IDC_RADIO_UP;
 		
-		select_ = SendMessage(GetDlgItem(hDlg, IDC_COMBO2), CB_GETCURSEL, 0, 0);
+		select_ = (int)(SendMessage(GetDlgItem(hDlg, IDC_COMBO2), CB_GETCURSEL, 0, 0));
 		return TRUE;
 	}
 	return FALSE;
